Add util package
